+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 2022-04-07 v4.0.0
+
+* fix: Syntax error in error message generator
+* feat: Partial sObjectTree API resource support - added CreateMultipleRecords to create multiple records in a single request
+    - see https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/dome_composite_sobject_tree_flat.htm
+* feat: updated ExternalIdInsertAndUpdate to return new InsertOrUpdateResponse object that adds the Created flag
+    - this includes a minor breaking change, as this method originally returned a CreateResponse object. Calls can be updated to use the InsertOrUpdateResponse object without any further breaking changes as it is inherited from the CreateResponse object.
+
 ### 2022-01-13 v3.1.0
 
 * feat: add .NET 6 support